+++ src-tauri/src/lib.rs
@@ -2455,6 +2455,7 @@ pub fn run() {
             browser_engine::browser_cdp_screenshot,
             browser_engine::browser_cdp_eval,
             browser_engine::browser_cdp_read_console,
+            browser_engine::browser_cdp_read_network,
             browser_engine::browser_cdp_wait_for,
             browser_engine::browser_cdp_assert,
             fetch_url,